change the csdsclient to interface{} to support other api versions
